Add files via upload
The aim of this deep analysis is to visualize a large “Potentially Excess Deaths from five leading causes of deaths” data set with several different rich visualization techniques. We have created some of the high-quality explanatory visualizations using Tableau and R that communicates a story from the data using Animation and interactive Dashboards.